road node placer window
To access this, go to the window in unity. Where it says sample node, place the road node prefabs in assets/prefabs, and where it says street place an empty which you use as a collection for a street. You will have to manually still move and place roadnodes, but it will automatically place a roadnode linked to the last roadnode, you can configure settings in the prefab to place them, they will all be parented, and roadnodes place at the location of the last node so you can place them pretty quick